requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
torch>=2.6.0
torchvision>=0.21.0
pillow>=10.0.0
python-multipart==0.0.9
numpy>=1.26.0
opencv-python>=4.10.0